+++ src/code/m_pause.c
@@ -1,6 +1,21 @@
-#include "global.h"
+#include "m_pause.h"
+#include "m_controller.h"
 
+void Pause_ct(Pause* pause){
+    pause->timer = 0;
+    pause->enabled = 0;
+}
 
-#pragma GLOBAL_ASM("asm/jp/nonmatchings/code/m_pause/Pause_ct.s")
+s32 Pause_proc(Pause* pause, Input* input){
+    if(CHECK_BTN_ALL(input->cur.button, R_TRIG) && CHECK_BTN_ALL(input->press.button, D_JPAD)){
+        pause->enabled = !pause->enabled;
+    }
 
-#pragma GLOBAL_ASM("asm/jp/nonmatchings/code/m_pause/Pause_proc.s")
+    if((!pause->enabled) || (CHECK_BTN_ALL(input->cur.button, Z_TRIG) && 
+        (CHECK_BTN_ALL(input->press.button, R_TRIG) || (CHECK_BTN_ALL(input->cur.button, R_TRIG) 
+        && (++pause->timer >= 9))))){
+            pause->timer = 0;
+            return 1;
+        }
+    return 0;
+}
